+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Track when a span was received. ([#2688](https://github.com/getsentry/relay/pull/2688))
 - Add context for NEL (Network Error Logging) reports to the event schema. ([#2421](https://github.com/getsentry/relay/pull/2421))
 - Add `validate_pii_selector` to CABI for safe fields validation. ([#2687](https://github.com/getsentry/relay/pull/2687))
+- Do not scrub Prisma spans. ([#2711](https://github.com/getsentry/relay/pull/2711))
 
 **Bug Fixes**: